# api.yaml
openapi: 3.0.1
info:
  title: Eventity API
  version: v1
paths:
...
  /api/users/me:
    get:
      tags:
        - User
      security:
      - Bearer: []
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        500:
          description: Server Error
  /api/users/{id}:
    get:
      tags:
        - User
      security:
      - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        500:
          description: Server Error
    put:
      tags:
        - User
      security:
      - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequestDto"
          text/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequestDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequestDto"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDto"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        500:
          description: Server Error
    delete:
      tags:
        - User
      security:
      - Bearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Success
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        500:
          description: Server Error
  /api/users:
    get:
      tags:
        - User
      security:
      - Bearer: []
      parameters:
        - name: login
          in: query
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponseDto"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        403:
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        500:
          description: Server Error
components:
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        message:
          type: string
          nullable: true
        details:
          type: string
          nullable: true
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: {}
...
    UpdateUserRequestDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        login:
          type: string
          minLength: 3
          maxLength: 20
          nullable: true
        password:
          type: string
          minLength: 6
          maxLength: 100
          nullable: true
      additionalProperties: false
...
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        login:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/UserRoleEnum"
      additionalProperties: false
    UserRoleEnum:
      type: integer
      enum:
        - User
        - Admin
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT Bearer token **_only_**
security:
  - Bearer: []